setup train and eval folders
