+++ .gitignore
@@ -1,1 +1,3 @@
 *.csv
+*.ps1
+**/__pycache__/
