+++ project/scripts/tags
@@ -4,4 +4,3 @@
 !_TAG_PROGRAM_NAME	Exuberant Ctags	//
 !_TAG_PROGRAM_URL	http://ctags.sourceforge.net	/official site/
 !_TAG_PROGRAM_VERSION	5.7	//
-main	.\run_evaluation.py	/^def main():$/;"	f
