+++ project/src/train/tags
@@ -1,0 +1,18 @@
+!_TAG_FILE_FORMAT	2	/extended format; --format=1 will not append ;" to lines/
+!_TAG_FILE_SORTED	1	/0=unsorted, 1=sorted, 2=foldcase/
+!_TAG_PROGRAM_AUTHOR	Darren Hiebert	/dhiebert@users.sourceforge.net/
+!_TAG_PROGRAM_NAME	Exuberant Ctags	//
+!_TAG_PROGRAM_URL	http://ctags.sourceforge.net	/official site/
+!_TAG_PROGRAM_VERSION	5.7	//
+calculate_feature_indices	.\index.py	/^def calculate_feature_indices(X, y):$/;"	f
+calculate_id	.\anfis.py	/^def calculate_id(indices, params):$/;"	f
+combine_file	.\model_pipeline.py	/^def combine_file(subject, model):$/;"	f
+gaussian_membership	.\anfis.py	/^def gaussian_membership(x, c, s):$/;"	f
+ification_report	.\model_pipeline.py	/^    report = classification_report(y_test_binary, y_pred)$/;"	c	function:optimizar
+ification_report	.\model_pipeline.py	/^from sklearn.metrics import classification_report, accuracy_score, confusion_matrix, roc_curve, auc, mean_squared_error, mutual_info_score$/;"	c
+ifiers	.\model_pipeline.py	/^    classifiers = {$/;"	c	function:train_pipeline
+ifiers	.\model_pipeline.py	/^    for name, clf in classifiers.items():$/;"	c	function:train_pipeline
+objective_function_with_progress	.\model_pipeline.py	/^def objective_function_with_progress(params):$/;"	f
+optimizar	.\model_pipeline.py	/^def optimizar(name,clf):$/;"	f
+show_result	.\output.py	/^def show_result(name,result):$/;"	f
+train_pipeline	.\model_pipeline.py	/^def train_pipeline(model):$/;"	f
